use a nature scenes api instead
